pyproject.toml
[project]
name = "taptapfirstproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "httpx>=0.28.1",
    "nltk>=3.9.1",
    "passlib>=1.7.4",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
